requirements.txt
# Azure Functions
azure-functions==1.21.3

# Azure SDK
azure-servicebus==7.11.4
azure-keyvault-secrets==4.7.0
azure-identity==1.15.0

# Meteorológicas
feedparser==6.0.11
requests==2.31.0
beautifulsoup4==4.12.3

# Geodata
geopandas==0.14.3
pandas==2.2.1
shapely==2.0.3
pyshp==2.3.1

# Twilio
twilio==9.2.3
